Postgres fixes for workflows interface
